Cluster improvements

+++ src/ingress/http-ingress.js
@@ -1,12 +1,14 @@
+import assert from 'assert/strict';
 import http, { Agent } from 'http';
 import net from 'net';
 import NodeCache from 'node-cache';
-import EventBus from '../eventbus/index.js';
+import EventBus from '../cluster/eventbus.js';
 import Listener from '../listener/index.js';
 import IngressUtils from './utils.js';
 import { Logger } from '../logger.js';
 import TunnelService from '../tunnel/tunnel-service.js';
 import AltNameService from './altname-service.js';
+import Node from '../cluster/cluster-node.js';
 import { ERROR_TUNNEL_NOT_FOUND,
          ERROR_TUNNEL_NOT_CONNECTED,
          ERROR_TUNNEL_HTTP_INGRESS_DISABLED,
@@ -16,7 +18,6 @@ import { ERROR_TUNNEL_NOT_FOUND,
          ERROR_TUNNEL_TARGET_CON_FAILED,
          ERROR_UNKNOWN_ERROR,
 } from '../utils/errors.js';
-import Node from '../utils/node.js';
 import {
     HTTP_HEADER_EXPOSR_VIA,
     HTTP_HEADER_X_FORWARDED_HOST,
@@ -40,7 +41,8 @@ class HttpIngress {
 
         this.destroyed = false;
         this.altNameService = new AltNameService();
-        this.tunnelService = new TunnelService();
+        this.tunnelService = opts.tunnelService;
+        assert(this.tunnelService instanceof TunnelService);
         this.httpListener = Listener.acquire('http', opts.port);
         this._requestHandler = this.httpListener.use('request', { logger: this.logger, prio: 1 }, async (ctx, next) => {
             if (!await this.handleRequest(ctx.req, ctx.res, ctx.baseUrl)) {
@@ -414,13 +416,15 @@ class HttpIngress {
     }
 
     async destroy() {
+        if (this.destroyed) {
+            return;
+        }
         this.destroyed = true;
         this.httpListener.removeHandler('request', this._requestHandler);
         this.httpListener.removeHandler('upgrade', this._upgradeHandler);
         return Promise.allSettled([
             this.altNameService.destroy(),
             this.eventBus.destroy(),
-            this.tunnelService.destroy(),
             Listener.release('http', this.opts.port),
         ]);
     }
